+++ lib/daisy_ui_components/textarea.ex
@@ -12,7 +12,7 @@ defmodule DaisyUIComponents.Textarea do
   attr :color, :string, values: colors()
   attr :ghost, :boolean, default: false
   attr :size, :string, values: sizes()
-  attr :rest, :global, include: ~w(name)
+  attr :rest, :global, include: ~w(name rows cols)
   slot :inner_block
 
   def textarea(assigns) do